requirements.txt
Flask==2.0.2
PyPDF2==2.10.3
langchain==0.0.24
langchain_google_genai==0.0.24
google.generativeai==0.0.24
dotenv==0.20.0